Added histogram array  & a invalidCard array to SpireDeck.java
Added a histogram array that increments the data when the cost is already verified
and added to the total cost. InvalidCard array only adds card when they have a invalid
energy value. When the array reaches 10 objects, the program stops. A total card count was
also added and stops the program if it gets above 1000